terminada e testada implementação de classes e interfaces

+++ src/servicos/ServicoPagamentoOnline.java
@@ -3,5 +3,5 @@ package servicos;
 public interface ServicoPagamentoOnline {
 
     public double taxaPagamento(double valor);
-    public double juro(double valor, double mes);
+    public double juro(double valor, int mes);
 }
